refactor: split basic `DefaultSelectionManager` modules
