+++ packages/blocks/src/page-block/default/selection-manager/index.ts
@@ -1,0 +1,5 @@
+export { DefaultSelectionManager } from './default-selection-manager.js';
+export {
+  PageSelectionState,
+  type PageSelectionType,
+} from './selection-state.js';
