+++ packages/blocks/src/page-block/utils/bind-hotkey.ts
@@ -27,7 +27,7 @@ import {
   Point,
 } from '../../__internal__/utils/index.js';
 import type { DefaultPageSlots } from '../default/default-page-block.js';
-import type { DefaultSelectionManager } from '../default/selection-manager.js';
+import type { DefaultSelectionManager } from '../default/selection-manager/index.js';
 import { handleSelectAll } from '../utils/index.js';
 import { formatConfig } from './const.js';
 import {
